+++ test/with-typescript/src/App.tsx
@@ -339,6 +339,14 @@ function getEmailPasswordConfigs() {
         signInAndUpFeature: {
             signInForm: {
                 style: theme.style,
+                formFields: [
+                    {
+                        id: "email",
+                        label: "Email",
+                        nonOptionalErrorMsg: "Please add your email",
+                        getDefaultValue: () => "abc@xyz.com",
+                    },
+                ],
             },
             signUpForm: {
                 style: theme.style,
@@ -371,6 +379,72 @@ function getEmailPasswordConfigs() {
                         id: "country",
                         label: "Your Country",
                         placeholder: "Where do you live?",
+                        optional: true,
+                    },
+                    {
+                        id: "terms",
+                        label: "",
+                        optional: false,
+                        nonOptionalErrorMsg: "Please check Terms and conditions",
+                        getDefaultValue: () => "true",
+                        inputComponent: (inputProps) => (
+                            <div
+                                style={{
+                                    display: "flex",
+                                    alignItems: "center",
+                                    justifyContent: "left",
+                                }}>
+                                <input
+                                    name={inputProps.name}
+                                    type="checkbox"
+                                    value={inputProps.value}
+                                    checked={inputProps.value === "true"}
+                                    onChange={(e) => {
+                                        if (inputProps.onChange) {
+                                            inputProps.onChange(e.target.checked.toString());
+                                        }
+                                    }}></input>
+                                <span style={{ marginLeft: 5 }}>I agree to the terms and conditions</span>
+                            </div>
+                        ),
+                        validate: async (value) => {
+                            if (value === "true") {
+                                return undefined;
+                            }
+                            return "Please check Terms and conditions";
+                        },
+                    },
+                    {
+                        id: "select",
+                        label: "Select",
+                        getDefaultValue: () => "option 2",
+                        inputComponent: (inputProps) => (
+                            <select
+                                onBlur={(e) => {
+                                    if (inputProps.onInputBlur) {
+                                        inputProps.onInputBlur(e.target.value);
+                                    }
+                                }}
+                                onFocus={(e) => {
+                                    if (inputProps.onInputFocus) {
+                                        inputProps.onInputFocus(e.target.value);
+                                    }
+                                }}
+                                value={inputProps.value}
+                                name={inputProps.name}
+                                onChange={(e) => {
+                                    if (inputProps.onChange) {
+                                        inputProps.onChange(e.target.value);
+                                    }
+                                }}>
+                                <option value="" disabled hidden>
+                                    Select an option
+                                </option>
+                                <option value="option 1">Option 1</option>
+                                <option value="option 2">Option 2</option>
+                                <option value="option 3">Option 3</option>
+                            </select>
+                        ),
                         optional: true,
                     },
                 ],
@@ -487,9 +561,115 @@ function getThirdPartyEmailPasswordConfigs() {
         },
         signInAndUpFeature: {
             style: theme.style,
+            signInForm: {
+                formFields: [
+                    {
+                        id: "email",
+                        label: "Email",
+                        nonOptionalErrorMsg: "Please add your email",
+                        getDefaultValue: () => "abc@xyz.com",
+                    },
+                ],
+            },
             signUpForm: {
                 privacyPolicyLink: "https://supertokens.io/legal/privacy-policy",
                 termsOfServiceLink: "https://supertokens.io/legal/terms-and-conditions",
+                formFields: [
+                    {
+                        id: "email",
+                        label: "Your Email",
+                        placeholder: "Your work email",
+                    },
+                    {
+                        id: "name",
+                        label: "Full name",
+                        placeholder: "First name and last name",
+                    },
+                    {
+                        id: "age",
+                        label: "Your age",
+                        placeholder: "How old are you?",
+                        validate: async (value) => {
+                            if (parseInt(value) > 18) {
+                                return undefined;
+                            }
+
+                            return "You must be over 18 to register";
+                        },
+                    },
+                    {
+                        id: "country",
+                        label: "Your Country",
+                        placeholder: "Where do you live?",
+                        optional: true,
+                    },
+                    {
+                        id: "terms",
+                        label: "",
+                        optional: false,
+                        nonOptionalErrorMsg: "Please check Terms and conditions",
+                        getDefaultValue: () => "true",
+                        inputComponent: (inputProps) => (
+                            <div
+                                style={{
+                                    display: "flex",
+                                    alignItems: "center",
+                                    justifyContent: "left",
+                                }}>
+                                <input
+                                    name={inputProps.name}
+                                    type="checkbox"
+                                    value={inputProps.value}
+                                    checked={inputProps.value === "true"}
+                                    onChange={(e) => {
+                                        if (inputProps.onChange) {
+                                            inputProps.onChange(e.target.checked.toString());
+                                        }
+                                    }}></input>
+                                <span style={{ marginLeft: 5 }}>I agree to the terms and conditions</span>
+                            </div>
+                        ),
+                        validate: async (value) => {
+                            if (value === "true") {
+                                return undefined;
+                            }
+                            return "Please check Terms and conditions";
+                        },
+                    },
+                    {
+                        id: "select",
+                        label: "Select",
+                        getDefaultValue: () => "option 2",
+                        inputComponent: (inputProps) => (
+                            <select
+                                onBlur={(e) => {
+                                    if (inputProps.onInputBlur) {
+                                        inputProps.onInputBlur(e.target.value);
+                                    }
+                                }}
+                                onFocus={(e) => {
+                                    if (inputProps.onInputFocus) {
+                                        inputProps.onInputFocus(e.target.value);
+                                    }
+                                }}
+                                value={inputProps.value}
+                                name={inputProps.name}
+                                onChange={(e) => {
+                                    if (inputProps.onChange) {
+                                        inputProps.onChange(e.target.value);
+                                    }
+                                }}>
+                                <option value="" disabled hidden>
+                                    Select an option
+                                </option>
+                                <option value="option 1">Option 1</option>
+                                <option value="option 2">Option 2</option>
+                                <option value="option 3">Option 3</option>
+                            </select>
+                        ),
+                        optional: true,
+                    },
+                ],
             },
             providers: [
                 ThirdPartyEmailPassword.Github.init(),